requirement.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
djangorestframework==3.13.1
idna==3.4
mysqlclient==2.1.1
Pillow==9.2.0
pycodestyle==2.9.1
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
